Add a flag whether to reuse containers
Since containers are easy and cheap to start and destroy, I find it more
convenient to be certain any job will start in a fresh container. Any
residue from a previous job (might be from a different team) should be
gone completely.

For this, it is not sufficient to change "Idle termination time" to `0`,
since queued jobs will still be executed on used containers.

Instead, I added a flag "Reusable", which will in effect suspend a
"computer" as soon as a job is scheduled on it. This means the computer
will remain idle after the job is finished, and therefore the "Idle
termination time" (still `0` obviously) will take effect. In practice
every job will now run on a fresh container.